Fix map node visibility
When we built for the device, we couldn't see the map markers, because the
application couldn't find their image files. I had to place the image
files in phonegap/www/img in order to reference them and gain access to
them on the device. Note: however that the web app can no longer find the
images now.