Fix header text on views
I was able to fix the header text on the views. As I mentioned in tfs, we
won't be able to correctly test the headers for the info page and each
spcific page until we rebase to a version of the app that contains
education information. Also, we will have to deal with the list page
population bug fix when it comes to searching for a product in order to
test the list page header and see if it correctly works.