Add disclosures to list items
We added the disclosure arrow buttons to the list page list items. These
disclosures make it so that a user can click once in order to go to a
details page, instead of a double tap or a highlight and navigate tap. To
get this to work I updated some of the structure of the controller
functions so that vars passed into the functions were instantiated in a
similar maner, so that we could easily follow the transfer of data from
the view to the controller.